Add mob management function and help message
My team would like to be able to manage who is included
in the mob more proactively than waiting for a rotation
to enable find_next.go to determine who is the next
typist. This change stubs out the entrypoint for the
management of the participants in the session. Further
changes are intended to include a method for storing
participant names in a temporary file which will be
deleted as part of the `mob done` command.